use fetch instead of require in api_controller_spec tests
For strong parameter nonsense. Using fetch allows an empty hash instead
of failing outright.
We're getting a test failure on TravisCI related to strong parameters on
an index GET call. I can't replicate it locally, but this is a change
that should happen on the controller related to the failure.