default API configuration for generated doc examples
So the problem is as it stands that the only place you configure
versions in the server is in the routes, so there's no easy way to link
models to actual URLs without some mad introspection. Since we can't
guarentee what form the versions will take we can't find the routes.
So this commit makes the generated docs use a default namespace and
version ('API' and 'v1' respectively). To get the real namespace, the
dev will need to load and configure their client side API on the server
or set them directly on Daylight::Documentation in an initializer.

Updated docs.